Added object check before adding raw response property
If soap method does not return an object the existing implementation fails because trying to add a prototype/field to a value type (i.e. string).   So that an existing soap user can wrap with soap-as-promised ok, I've kept the return value as-is rather than creating a wrapper object (like with null values).